requirements.txt
openai>=1.0.0
anthropic>=0.20.0
gigachat>=0.1.9
pandas>=2.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0